aio: for async operations, make the iter argument persistent
When implementing async read/write operations, the complexity of having
to duplicate the iter argument before passing to another thread leads to
duplicate code.  There is no reason async operations issued by the aio
core need to be placed on the stack when an aio_kiocb is allocated for
each operation, so put the iter and iovec into aio_kiocb instead of on
the stack.

Signed-off-by: Benjamin LaHaise <ben.lahaise@solacesystems.com>
Signed-off-by: Benjamin LaHaise <bcrl@kvack.org>