m68k: provide __phys_to_pfn() and __pfn_to_phys()
The defconfig build of m68k was failing with the error: implicit
declaration of function '__pfn_to_phys'

Other architectures have added <asm/memory.h>, but if we do so here then
we will also get redeclaration of some other functions.  So it is better
to copy these macros into page.h.

[geert: apply to page.h instead of page_mm.h to cover nommu, reword]
Fixes: 0a3c3bf11240 ("x86, mm: introduce vmem_altmap to augment vmemmap_populate()")
Signed-off-by: Sudip Mukherjee <sudip@vectorindia.org>
Signed-off-by: Geert Uytterhoeven <geert@linux-m68k.org>
Reported-by: Guenter Roeck <linux@roeck-us.net> (m68knommu)
Cc: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>