dax: remove unnecessary rechecking of i_size
When i_mmap_lock (or the page lock) was the only protection against
truncate, we checked i_size at the beginning of the fault handler, then
rechecked it after acquiring the lock.  Since the fliesystems now exclude
truncate from racing with the fault handler, we no longer need to recheck
i_size.  We do, of course, still need to check i_size at the entry to the
fault handler.

Also remove the now-unnecessary acquisitions of i_mmap_lock.  One of the
acquisitions is still needed, so put a big fat comment beside it to
prevent the well-intentioned from removing it.

Signed-off-by: Matthew Wilcox <matthew.r.wilcox@intel.com>
Cc: "Kirill A. Shutemov" <kirill@shutemov.name>
Cc: Ross Zwisler <ross.zwisler@linux.intel.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>