IB/IPoIB: Allow setting the device address
In IPoIB/rdmacm networks, the device address of an IPoIB
interface is used to exchange information between nodes
needed for communication.

Currently an IPoIB interface will always be created with a
constant device address based on its node GUID.

This change adds the ability to set the device address of an IPoIB
interface by value. The mac address ndo is used for that.

The control flow can be divided to two parts:
1) The GID value is already in the GID table,
   in this case the interface will be able to set carrier up.

2) The GID value is not yet in the GID table,
   in this case the interface won't try to join the multicast group
   and will wait (listen on GID_CHANGE event) until the GID is inserted.

In order to track those changes, we will add a new flag:
* IPOIB_FLAG_DEV_ADDR_SET.

When this flag is set, it will indicate that the dev_addr is a
based on a value in the GID table. The dev_addr change will
trigger the bit clear operation.

Per IB specification the port GUID can't be changed if the
module is loaded. Port GUID is the basis for GID at index 0
which is the basis for the default device address of a IPoIB
interface.

There are devices that don't follow the spec and they change
the port GUID while HCA is powered on, so in order do not break
existing user space applications, the following flow is suggested.

We will check if the user wants to control the device address
and we will assume that he no longer wishes to control it, if
he sets the device address back to be based on GID index 0.

In order to track this, we will add an additional flag:
* IPOIB_FLAG_DEV_ADDR_CTRL

When setting the device address, there is lack of validation for
12 upper bytes of the device address (flags, qpn, subnet prefix)
as those bytes are not under the control of the user.

Signed-off-by: Mark Bloch <markb@mellanox.com>
Signed-off-by: Leon Romanovsky <leon@kernel.org>