ext4: optimize group search for inode allocation
Added a check at the start of group search loop to
avoid looping unecessarily in case of empty group.
This also allow group search to jump directly to
"found_flex_bg" with "stats" and "group" already set,
so there is no need to go through the extra steps of
setting "best_desc" , "best_group" and then break
out of the loop just to set "stats" and "group" again.

Reviewed-by: Andreas Dilger <adilger@dilger.ca>
Signed-off-by: Lokesh Nagappa Jaliminche <lokesh.jaliminche@gmail.com>
Signed-off-by: Theodore Ts'o <tytso@mit.edu>