aio: add support for in-submit openat
Using the LOOKUP_RCU infrastructure added for open(), implement such
functionality to enable in io_submit() openat() that does a non-blocking
file open operation.  This avoids the overhead of punting to another
kernel thread to complete the open operation when the files and data are
already in the dcache.  This helps cut simple aio openat() from ~60-90K
cycles to ~24-45K cycles on my test system.

Signed-off-by: Benjamin LaHaise <bcrl@kvack.org>