dm raid1: do not error successfully written bios during suspend
The "mirror" target was incorrectly returning -EIO, in the
suspend-with-flush case of hold_bio(), for bios that had been written
correctly.

Errors are already handled properly for a device that is itself faulty
(all mirror legs have failed).  So hold_bio() is modified to reflect
that relevant faults are handled elsewhere -- and in the case of
suspend-with-flush no error should be returned.

Also, fix comment in do_failures() to reflect this change.

Fixes: 60f355ead ("dm raid1: hold write bios when errors are handled")
Signed-off-by: Heinz Mauelshagen <heinzm@redhat.com>
Signed-off-by: Mike Snitzer <snitzer@redhat.com>
Cc: stable@vger.kenel.org