futex: Add support for attached futexes
The standard futex mechanism in the Linux kernel uses a global hash to store
transient state. Collisions on that hash can lead to performance degradation
and on real-time enabled kernels even to priority inversions.

To guarantee futexes without collisions on the global kernel hash, we provide
a mechanism to attach to a futex. This creates futex private state which
avoids hash collisions and on NUMA systems also cross node memory access.

To utilize this mechanism each thread has to attach to the futex before any
other operations on that futex.

The inner workings are as follows:

Attach:

    sys_futex(FUTEX_ATTACH | FUTEX_ATTACHED, uaddr, ....);

    If this is the first attach to uaddr then a 'global state' object is
    created. This global state contains a futex hash bucket and a futex_q
    object which is enqueued into the global hash for reference so subsequent
    attachers can find it. Each attacher takes a reference count on the
    'global state' object and hashes 'uaddr' into a thread local hash. This
    thread local hash is lock free and dynamically expanded to avoid
    collisions. Each populated entry in the thread local hash stores 'uaddr'
    and a pointer to the 'global state' object.

Futex ops:

    sys_futex(FUTEX_XXX | FUTEX_ATTACHED, uaddr, ....);

    If the attached flag is set, then 'uaddr' is hashed and the thread local
    hash is checked whether the hash entry contains 'uaddr'. If no, an error
    code is returned. If yes, the hash slot number is stored in the futex key
    which is used for further operations on the futex. When the hash bucket is
    looked up then attached futexes will use the slot number to retrieve the
    pointer to the 'global state' object and use the embedded hash bucket for
    the operation. Non-attached futexes just use the global hash as before.

Detach:

    sys_futex(FUTEX_DETACH | FUTEX_ATTACHED, uaddr, ....);
   
    Detach removes the entry in the thread local hash and decrements the
    refcount on the 'global state' object. Once the refcount drops to zero the
    'global state' object is removed from the global hash and destroyed.

    Thread exit cleans up the thread local hash and the 'global state' objects
    as we do for other futex related storage already.

The thread local hash and the 'global state' object are allocated on the node
on which the attaching thread runs.

Attached mode works with all futex operations and with both private and shared
futexes. For operations which involve two futexes, i.e. FUTEX_REQUEUE_* both
futexes have to be either attached or detached (like FUTEX_PRIVATE).

Why not auto attaching?

    Auto attaching has the following problems:

     - Memory consumption
     - Life time issues
     - Performance issues due to the necessary allocations

    So, no. It must be opt-in and reserved for explicit isolation purposes.

A modified version of 'perf bench futex hash' shows the following results:

 # ./perf bench futex hash -s -r 60
 # Running 'futex/hash' benchmark:
 Run summary [PID 4456]: 144 threads, each operating on 1024 [private ] futexes for 60 secs.

 Averaged 1451441 operations/sec (+- 3.65%), total secs = 60

 Perf stat analysis for such a run:

 Performance counter stats for './perf bench futex hash -s -r 60':

     9,400,551,930      node-load-misses
        74,404,109      node-loads
    11,001,148,922      node-store-misses
       574,505,561      node-stores

      60.016993846 seconds time elapsed

 # Running 'futex/hash' benchmark:
 Run summary [PID 3675]: 144 threads, each operating on 1024 [private attached] futexes for 60 secs.

 Averaged 1709712 operations/sec (+- 4.67%), total secs = 60

That's a performance increase of 18%.

 Perf stat analysis for such a run:

 Performance counter stats for './perf bench futex hash -s -r 60 -a':

     2,609,871,190      node-load-misses
         9,671,312      node-loads
     2,346,187,670      node-store-misses
       144,453,796      node-stores

      60.020374008 seconds time elapsed

Now someone might argue that this is an unrealistic test case. Sure the number
of futexes here exceeds the size of the global hash:

   144 * 1024  >  144 * 256

But from real world applications with a way smaller number of futexes we have
observed hash collisions on real-time and non real-time systems, which cause
performance or determinism issues. Surely we could increase the hash size, but
that just reduces the probability and does not exclude the chance to hit a bad
spot.

We wrote a - too ugly to share - collision scanner which runs on each node of
a NUMA system. It scans through malloced and mmaped memory and invokes
thousands of consecutive (failing like the hash bench test) futex_wait
operations on each address and observes the consumed time. In case of a
collision the time jumps massively due to the atomic_inc/dec and the spinlock
operation on the hash bucket and the resulting cache line bouncing.

Once it detected a collision it stays there for 60 seconds and then switches
to attached mode. This immediately brings back the performance on the involved
scanners. A collision with two threads and therefor two futexes on different
nodes results in a performance degradation of 30 and more percent in this
test. On real-time enabled systems even a one time collision can have fatal
consequences due to possibly unbound priority inversions.

An interesting observation is that the time to find a collision varies from a
few seconds (20 addresses scanned) to a couple of hours. That variation is due
to the different mm_struct addresses which are mixed into the global hash and
different malloc/mmap addresses which are handed out on each run.

A few smallish issues:

  - we must limit the number of allocated 'global state' objects. That must
    be a per process limit, but I have no idea where to put it best and how
    we let the sysadmin tune it.
  
  - this needs massive review and exposure to fuzzers as I don't want to end
    up with another exploit disaster in that code.

[ tglx: Wrote changelog ]

Suggested-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>