futex: Make key init a helper function
Support for attached futexes requires to store information in the futex
key. Make the key init a helper function and let it return the futex uaddr.

At the call sites check the returned uaddr for NULL so the attached mode can
abort the operation when it detects that an attached operation is attempted on
a non attached futex.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>