huge tmpfs: get_unmapped_area align & fault supply huge page
Now make the shmem.c changes necessary for mapping its huge pages into
userspace with huge pmds: without actually doing so, since that needs
changes in huge_memory.c and across mm, better left to another patch.

Provide a shmem_get_unmapped_area method in file_operations, called at
mmap time to decide the mapping address.  It could be conditional on
CONFIG_TRANSPARENT_HUGEPAGE, but save #ifdefs in other places by making it
unconditional.

shmem_get_unmapped_area() first calls the usual mm->get_unmapped_area
(which we treat as a black box, highly dependent on architecture and
config and executable layout).  Lots of conditions, and in most cases it
just goes with the address that chose; but when our huge stars are rightly
aligned, yet that did not provide a suitable address, go back to ask for a
larger arena, within which to align the mapping suitably.

There have to be some direct calls to shmem_get_unmapped_area(), not via
the file_operations: because of the way shmem_zero_setup() is called to
create a shmem object late in the mmap sequence, when MAP_SHARED is
requested with MAP_ANONYMOUS or /dev/zero.  Though this only matters when
/proc/sys/vm/shmem_huge has been set.

Then at fault time, shmem_fault() does its usual shmem_getpage_gfp(), and
if caller __do_fault() passed FAULT_FLAG_MAY_HUGE (in later patch), checks
if the 4kB page returned is PageTeam, and, subject to further conditions,
proceeds to populate the whole of the huge page (if it was not already
fully populated and uptodate: use PG_owner_priv_1 PageChecked to save
repeating all this each time the object is mapped); then returns it to
__do_fault() with a VM_FAULT_HUGE flag to request a huge pmd.

Two conditions you might expect, which are not enforced.  Originally I
intended to support just MAP_SHARED at this stage, which should be good
enough for a first implementation; but support for MAP_PRIVATE (on read
fault) needs so little further change, that it was well worth supporting
too - it opens up the opportunity to copy your x86_64 ELF executables to
huge tmpfs, their text then automatically mapped huge.

The other missing condition: shmem_getpage_gfp() is checking that the
fault falls within (4kB-rounded-up) i_size, but shmem_fault() maps hugely
even when the tail of the 2MB falls outside the (4kB-rounded-up) i_size.
This is intentional, but may need reconsideration - especially in the
MAP_PRIVATE case (is it right for a private mapping to allocate "hidden"
pages to the object beyond its EOF?).  The intent is that an application
can indicate its desire for huge pmds throughout, even of the tail, by
using a hugely-rounded-up mmap size; but we might end up retracting this,
asking for fallocate to be used explicitly for that.  (hugetlbfs behaves
even less standardly: its mmap extends the i_size of the object.)

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>