huge tmpfs: mem_cgroup shmem_pmdmapped accounting
Grep now for shmem_pmdmapped in memory.stat (and also for "total_..." in a
hierarchical setting).

This metric allows for easy checking on a per-cgroup basis of the amount
of page team memory hugely mapped (at least once) out there.

The metric is counted towards the cgroup owning the page (unlike in an
event such as THP split) because the team page may be mapped hugely for
the first time via a shared map in some other process.

Moved up mem_group_move_account()'s PageWriteback block: that movement is
irrelevant to this patch, but lets us concentrate better on the PageTeam
locking issues which follow in the next patch.

Signed-off-by: Andres Lagar-Cavilla <andreslc@google.com>
Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>