huge tmpfs recovery: framework for reconstituting huge pages
Huge tmpfs is great when you're allocated a huge page from the start; but
not much use if there was a shortage of huge pages at that time, or your
huge pages were disbanded and swapped out under pressure, and now paged
back in 4k pieces once the pressure has eased.  At present the best you
can do is copy your original file, and start afresh on the unfragmented
copy; but we do need a better answer.

The approach taken here is driven from page fault: assembling a huge page
from existing pieces is more expensive than initial allocation from an
empty huge page, and the work done quite likely to be wasted, unless
there's some evidence that a huge TLB mapping will be useful to the
process.  A page fault in a suitable area suggests that it may.

So we adjust the original "Shall we map a huge page hugely?" tests in
shmem_fault(), to distinguish what can be done on this occasion from what
may be possible later: invoking shmem_huge_recovery() when we cannot map a
huge page now, but might be able to use one later.

It's likely that this is over-eager, that it needs some rate-limiting, and
should be tuned by the number of faults which occur in the extent.  Such
information will have to be stored somewhere: probably in the extent's
recovery work struct; but no attempt to do so in this series.

So as not to add latency to the fault, shmem_huge_recovery() just enqueues
a work item - with no consideration for all the flavors of workqueue that
might be used: would something special be better?

But skips it if this range of the file is already on the queue (which is
both more efficient, and avoids awkward races later), or if too many items
are currently enqueued.  "Too many" defaults to more than 8, tunable via
/proc/sys/vm/shmem_huge_recoveries - seems more appropriate than adding it
into the huge=N mount option.  Why 8?  Well, anon THP's khugepaged is
equivalent to 1, but work queues let us be less restrictive.  Initializing
or tuning it to 0 completely disables huge tmpfs recovery.

shmem_recovery_work() is where the huge page is allocated - using
__alloc_pages_node() rather than alloc_pages_vma(), like anon THP does
nowadays: ignoring vma mempol complications for now, though I'm sure our
NUMA behavior here will need to be improved very soon.  Population and
remap phases left as stubs in this framework commit.

But a fresh huge page is not necessarily allocated: page migration is
never sure to succeed, so it's wiser to allow a work item to resume on a
huge page begun by an earlier, than re-migrate all its pages so far
instantiated, to yet another huge page.  Sometimes an unfinished huge page
can be easily recognized by PageTeam; but sometimes it has to be located,
by the same SHMEM_TAG_HUGEHOLE mechanism that exposes it to the hugehole
shrinker.  Clear the tag to prevent the shrinker from interfering
(unexpectedly disbanding) while in shmem_populate_hugeteam() itself.

If shmem_huge_recoveries is enabled, shmem_alloc_page()'s retry after
shrinking is disabled: in early testing, the shrinker was too eager to
undo the work of recovery.  That was probably a side-effect of bugs at
that time, but it still seems right to reduce the latency of shmem_fault()
when it has a second chance.

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>