huge tmpfs recovery: shmem_recovery_populate to fill huge page
The outline of shmem_recovery_populate() is straightforward: a loop
applying shmem_getpage_gfp() to each offset that belongs to the extent,
converting swapcache to pagecache, checking existing pagecache, or
allocating new pagecache; adding each correctly placed page into the team,
or isolating each misplaced page, for passing to migrate_pages() at the
end of the loop.  Repeated (skipping quickly over those pages resolved in
the previous pass) to add in those pages just migrated, until the team is
complete, or cannot be completed.

But the details are difficult: not so much an architected design, as a
series of improvisations arrived at by trial and much error, which in the
end works well.  It has to cope with a variety of races: pages being
concurrently created or swapped in or disbanded or deleted by other
actors.

Most awkward is the handling of the head page of the team: which, as
usual, needs PageTeam and mapping and index set even before it is
instantiated (so that shmem_hugeteam_lookup() can validate tails against
it); but must not be confused with an instantiated team page until
PageSwapBacked is set.  This awkwardness is compounded by the (unlocked)
interval between when migrate_pages() migrates an old page into its new
team location, and the next repeat of the loop which fixes the new
location as PageTeam.  Yet migrate_page_move_mapping() will have already
composed PageSwapBacked (from old page) with PageTeam (from new) in the
case of the team head: "account_head" to track this case correctly, but a
later patch offers a tighter alternative to remove the need for it.

That interval between migration and enteamment also involves giving up on
the SHMEM_RETRY_HUGE_PAGE option from shmem_hugeteam_lookup():
SHMEM_ALLOC_SMALL_PAGE is not ideal, but does for now; and the later patch
can restore the SHMEM_RETRY_HUGE_PAGE optimization.

Note: this series was originally written with a swapin pass before
population, whereas this commit simply lets shmem_getpage_gfp() do the
swapin synchronously.  The swapin pass is reintroduced as an optimization
afterwards, but some comments on swap in this commit may anticipate that:
sorry, a precise sequence of developing comments took too much trouble.

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>