mm, page_alloc: move might_sleep_if check to the allocator slowpath
There is a debugging check for callers that specify __GFP_DIRECT_RECLAIM
from a context that cannot sleep.  Triggering this is almost certainly a
bug but it's also overhead in the fast path.  Move the check to the slow
path.  It'll be harder to trigger as it'll only be checked when watermarks
are depleted but it'll also only be checked in a path that can sleep.

Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Jesper Dangaard Brouer <brouer@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>