powerpc/mm/radix: Add MMU_FTR_RADIX
We are going to add asm changes in the follow up patches. Add the
feature bit now so that we can get it all build.

mpe: Note that if CONFIG_PPC_RADIX_MMU=n we define MMU_FTR_RADIX to
zero. This has the effect of turning all the radix_enabled() checks into
if (0), which the compiler is smart enough to elide completely.

Signed-off-by: Aneesh Kumar K.V <aneesh.kumar@linux.vnet.ibm.com>
Acked-by: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Michael Ellerman <mpe@ellerman.id.au>